+++ fabric/src/main/java/house/greenhouse/bovinesandbuttercups/client/BovinesAndButtercupsFabricClient.java
@@ -1,5 +1,6 @@
 package house.greenhouse.bovinesandbuttercups.client;
 
+import house.greenhouse.bovinesandbuttercups.client.renderer.entity.model.MoobloomModel;
 import net.fabricmc.api.ClientModInitializer;
 import net.fabricmc.fabric.api.blockrenderlayer.v1.BlockRenderLayerMap;
 import net.fabricmc.fabric.api.client.model.loading.v1.ModelLoadingPlugin;
@@ -58,7 +59,7 @@ public class BovinesAndButtercupsFabricClient implements ClientModInitializer {
         BovinesAndButtercupsClient.init(new BovinesClientHelperFabric());
         BovineBlockstateTypes.init();
 
-        EntityModelLayerRegistry.registerModelLayer(BovinesModelLayers.MOOBLOOM_MODEL_LAYER, CowModel::createBodyLayer);
+        EntityModelLayerRegistry.registerModelLayer(BovinesModelLayers.MOOBLOOM_MODEL_LAYER, MoobloomModel::createBodyLayer);
         EntityModelLayerRegistry.registerModelLayer(BovinesModelLayers.FLOWER_CROWN_MODEL_LAYER, () -> FlowerCrownModel.createLayer(new CubeDeformation(0.5F)));
         EntityModelLayerRegistry.registerModelLayer(BovinesModelLayers.PIGLIN_FLOWER_CROWN_MODEL_LAYER, () -> FlowerCrownModel.createLayer(new CubeDeformation(1.5F, 0.5F, 0.5F)));
         EntityRendererRegistry.register(BovinesEntityTypes.MOOBLOOM, MoobloomRenderer::new);